+++ housekeeping-vacuum-tool/src/main/java/com/hotels/housekeeping/tool/vacuum/validate/ValidationException.java
@@ -1,0 +1,7 @@
+package com.hotels.housekeeping.tool.vacuum.validate;
+
+public class ValidationException extends RuntimeException {
+
+  private static final long serialVersionUID = 1L;
+
+}