+++ housekeeping-vacuum-tool/src/main/java/com/hotels/housekeeping/tool/vacuum/validate/ValidationFailure.java
@@ -1,0 +1,14 @@
+package com.hotels.housekeeping.tool.vacuum.validate;
+
+public interface ValidationFailure {
+
+  /**
+   * @return explanation about the failure
+   */
+  String getMessage();
+
+  /**
+   * @return 'database.table' name that is invalid and cannot be vacuumed
+   */
+  String getQualifiedTableName();
+}